Move functions from Store to exchangerUtils
